requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///D:/bld/aiohttp_1767524540067/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1767719033/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_async-timeout_1767290444/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_attrs_1764875182/work
backports.zstd @ file:///D:/bld/bld/rattler-build_backports.zstd_1767045016/work
Brotli @ file:///D:/bld/brotli-split_1764017006836/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy @ file:///D:/bld/contourpy_1744743084386/work
cssselect==1.4.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_cycler_1764466758/work
DataRecorder==3.6.2
datasets @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_datasets_1769125358/work
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1769745019/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dill @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_dill_1764517905/work
DownloadKit==2.0.7
DrissionPage==4.1.1.2
et_xmlfile==2.0.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fake-useragent==2.2.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1768022204021/work
fonttools @ file:///D:/bld/fonttools_1765632560086/work
frozenlist @ file:///D:/bld/frozenlist_1752167168686/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1761840104528/work
greenlet==3.3.1
h11 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h11_1767729720/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet @ file:///D:/bld/bld/rattler-build_hf-xet_1763772668/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1761225853865/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1765794027227/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
kiwisolver @ file:///D:/bld/bld/rattler-build_kiwisolver_1762488749/work
lxml==6.0.2
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mpmath==1.3.0
multidict @ file:///D:/bld/multidict_1765460325754/work
multiprocess @ file:///D:/bld/bld/rattler-build_multiprocess_1762474370/work
munkres==1.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy @ file:///D:/bld/numpy_1747544591878/work/dist/numpy-2.2.6-cp310-cp310-win_amd64.whl#sha256=af66576ce916c795ffa81be9620c8138a3fda6e1713b2bf7775a630e47eca3a8
openpyxl==3.1.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas @ file:///D:/bld/pandas_1764614482805/work
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///D:/bld/bld/rattler-build_pillow_1767353195/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
playwright==1.58.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache @ file:///D:/bld/propcache_1744524966467/work
psutil @ file:///D:/bld/bld/rattler-build_psutil_1769678167/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.0
pyee==13.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1769003998/work
PySide6==6.10.1
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
PyYAML @ file:///D:/bld/pyyaml_1758891559676/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
regex @ file:///D:/bld/regex_1768558583329/work
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
requests-file==3.0.1
safetensors @ file:///D:/bld/safetensors_1763569508301/work
scikit-learn @ file:///D:/bld/scikit-learn_1757432814257/work/dist/scikit_learn-1.7.2-cp310-cp310-win_amd64.whl#sha256=80780e7bac7c11947bc8f76f74cce4a729277e8fc2c67f76db5038e95de040c9
scipy @ file:///C:/bld/scipy-split_1739790833536/work/dist/scipy-1.15.2-cp310-cp310-win_amd64.whl#sha256=4748853a0c723cda1fa2bdead847ba7c2443bde9cec3436e3e7243cfa54b55ff
sentence-transformers==5.2.2
shiboken6==6.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1762941432086/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.14.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tldextract==5.3.1
tokenizers @ file:///D:/bld/tokenizers_1764694754082/work/bindings/python
torch==2.10.0+cu126
torchvision==0.25.0+cu126
tornado @ file:///D:/bld/tornado_1765836387919/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1768754954982/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1765719872007/work
unicodedata2 @ file:///D:/bld/unicodedata2_1763054693050/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1769858610910/work
websocket-client==1.9.0
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
xxhash @ file:///D:/bld/python-xxhash_1762516211084/work
yarl @ file:///D:/bld/yarl_1761336846407/work
